requirements.txt
alembic==1.17.0
annotated-types==0.7.0
anyio==4.11.0
APScheduler==3.11.0
bcrypt==4.0.1
click==8.3.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.119.1
greenlet==3.2.4
gunicorn==23.0.0
h11==0.16.0
idna==3.11
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mysql-connector-python==9.4.0
packaging==25.0
passlib==1.7.4
psutil==7.1.1
pyasn1==0.6.1
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
pyotp==2.9.0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
rich==14.2.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.48.0
toml==0.10.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
uvicorn==0.38.0
watchdog==6.0.0
